refactor(test): 테스트 결과 확인하기 쉽게 수정
InputWordTest -> AnswerTest로 위치 변경
